requirements.txt
sdnotify==0.3.2
python-dateutil==2.8.0
paho-mqtt==1.4.0